Factories

+++ database/migrations/2026_03_05_082121_create_products_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->text('description');
-            $table->decimal('price');
+            $table->float('price');
             $table->foreignId('category_id')->references('id')->on('categories');
             $table->timestamps();
         });
@@ -25,3 +25,4 @@ return new class extends Migration
         Schema::dropIfExists('products');
     }
 };
+
